+++ src/main/java/com/gruppe27/fellesprosjekt/common/messages/AuthMessage.java
@@ -1,4 +1,4 @@
-package com.gruppe27.fellesprosjekt.common;
+package com.gruppe27.fellesprosjekt.common.messages;
 
 public class AuthMessage {
     public enum Command {
