+++ src/main/java/com/gruppe27/fellesprosjekt/common/messages/ErrorMessage.java
@@ -1,4 +1,4 @@
-package com.gruppe27.fellesprosjekt.common;
+package com.gruppe27.fellesprosjekt.common.messages;
 
 public class ErrorMessage {
     private String message;
