+++ src/main/java/com/gruppe27/fellesprosjekt/server/controllers/AuthController.java
@@ -1,10 +1,10 @@
 package com.gruppe27.fellesprosjekt.server.controllers;
 
-import com.gruppe27.fellesprosjekt.common.AuthMessage;
-import com.gruppe27.fellesprosjekt.common.ErrorMessage;
-import com.gruppe27.fellesprosjekt.common.GeneralMessage;
+import com.gruppe27.fellesprosjekt.common.messages.AuthMessage;
+import com.gruppe27.fellesprosjekt.common.messages.ErrorMessage;
+import com.gruppe27.fellesprosjekt.common.messages.GeneralMessage;
 import com.gruppe27.fellesprosjekt.common.User;
-import com.gruppe27.fellesprosjekt.common.UserMessage;
+import com.gruppe27.fellesprosjekt.common.messages.UserMessage;
 import com.gruppe27.fellesprosjekt.server.CalendarConnection;
 
 import java.sql.Connection;
